requirements.txt
kagglehub
pandas
numpy
matplotlib
seaborn
scikit-learn
Pillow
tqdm
torch
torchvision
customtkinter
optuna
